Moved code for authorization into authentication.go
